plugin: sys_status: publish state msg after updating uas
Before this commit, the custom mode string published in the
state message was computed using the autopilot type from the
previous heartbeat message--*not* the autopilot type from the
current hearbeat message.

Normally that isn't a problem, but when running a GCS and mavros
concurrently, both connected to an FCU that routes mavlink packets
(such as APM), then this causes the custom mode to be computed
incorrectly, because the mode string for the GCS's hearbeat packet
will be computed using the FCU's autopilot type, and the mode string
for the FCU's heartbeat packet will be computed using the GCS's
autopilot type.