Fixed build dependency for msg creation
Using MATLAB and roboticsSupportPackages to read and convert a ROSbag to timeseries.
The messages i'm interested in are the /mavros/setpoint_raw/attitude.msg.

So using the rosgenmsg('/home/user/catkin_ws/mavros') gave the error that the std_msgs and geometry_msgs didn't exist. This is because in the Package.xml they are define as depend, not as build_depend.

After fixing this, everything worked fine.