Removed x_lat, y_long, z_alt from WP
These values seemed to be used due to the fact that double had
a greater resolution than float and doubles are used in the
ros msg. However they were only ever used for printing. Since
the int version of these messages has a greater resolution I
figure it is more useful to print the true value in the mavlink
message rather than the ros message value