Added a capabilities change cb queue
Plugins can now write functions that they add to the
capabilities_cb_vec. These functions will be called only when there is a
change to the capabilities themselves not whenever the known status of
the fcu_capabilities change.

These functions should have a parameter of type
mavlink::common::MAV_PROTOCOL_CAPABILITY which is essentially just a
uint64_t however being more opinionated is helpful when looking for what
the canonical enum names are in the mavlink project header files.